# WIP: `ndimage.generic_filter`: potential performance (20x~150x) and feature improvements