--- a/scipy/ndimage/_delegators.py
+++ b/scipy/ndimage/_delegators.py
@@ -288,6 +288,12 @@ def value_indices_signature(arr, *args, **kwds):
     return array_namespace(arr)
 
 
+def vectorized_filter_signature(
+    input, function, size=None, footprint=None, output=None, *args, **kwds
+):
+    return array_namespace(input, footprint, _skip_if_dtype(output))
+
+
 def watershed_ift_signature(input, markers, structure=None, output=None):
     return array_namespace(input, markers, structure, _skip_if_dtype(output))