… scipy/ndimage/_ni_docstrings.py
… scipy/ndimage/tests/test_filters.py