#### Reference issue
With some work, could close:
gh-20185 (needs implementation, pretty easy)
gh-18437 (needs testing)
gh-17546 (needs implementation, very easy)
gh-8916 (needs testing)
gh-5137 (needs implementation, probably easy)

#### What does this implement/fix?
`ndimage.generic_filter` is implemented in compiled code. This is performant when using a `LowLevelCallable`, but there are downsides: it is slow for Python callables and the code is a bit intimidating to work on, so it tends to get less attention.

This PR adds a `_generic_filter` which could be used by `ndimage.generic_filter` for vectorized, pure-Python callables. It is much faster than `ndimage.generic_filter`, and in fact, it can even be faster than the specialized filters. Using lightly-modified `median_filter` example code, for instance:

```python3
import numpy as np
from scipy.ndimage._filters import _generic_filter
from scipy import ndimage, datasets
ascent = datasets.ascent()
size = 21  # odd to ensure median is an element of the data
ref1 = ndimage.median_filter(ascent, size=size)
ref2 = ndimage.generic_filter(ascent, np.median, size=size)
res = _generic_filter(ascent, np.median, size=size)
np.testing.assert_equal(res, ref1)  # passes
np.testing.assert_equal(res, ref2)  # passes

%timeit ndimage.generic_filter(ascent, np.median, size=size)  # 6.11 s ± 44.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
%timeit ndimage.median_filter(ascent, size=size)  # 664 ms ± 692 μs per loop (mean ± std. dev. of 7 runs, 1 loop each)
%timeit _generic_filter(ascent, np.median, size=size)  # 517 ms ± 2.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

For small footprints, 100x speedup hasn't been uncommon in testing.
```python3
size = (3, 4)
# `_generic_filter` currently respects output type of `np.mean`, `ndimage.generic_filter` doesn't
ascent = ascent.astype(dtype=np.float32)
ref = ndimage.generic_filter(ascent, np.mean, size=size)  # 1.26 s ± 17.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
res = _generic_filter(ascent, np.mean, size=size)  # 8.03 ms ± 53.3 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
np.testing.assert_allclose(res, ref, rtol=1e-7)  # passes
```

Incidentally, we also get all dtypes for free. It would be quite easy to implement most of the features (`mask`, `mode='valid', higher-dimensional output) in the linked issues.

**Currently**, this is memory-intensive - like the existing filters, I think; see gh-4878. However, the only very memory-inefficient part is the execution of the callable, which seems to inflate the sliding window view to a full sized array during evaluation. The hope is that we can get decent performance without all the memory usage by passing the view to Cython and just doing the loop in compiled code. If we go that route, I can also imagine adding an interface for efficiently *updating* the filtered value (accounting only for the elements that are leaving and entering the footprint) rather than recomputing from scratch.

Thoughts on how/whether we can use this? 

#### Additional information
I think it's possible to add this in a backward compatible way, since it's pretty easy to write comprehensive, property-based tests:

<details>

```python3
import numpy as np
from scipy.ndimage import generic_filter
from scipy.ndimage._filters import _generic_filter

rng = np.random.default_rng(2394582934548265)
input = rng.random(size=(11, 12, 13))
for axes_size in [(None, (3, 4, 5)), ((0, 2), (3, 4)), ((-1,), (5,))]:
    for origin in [0, 1]:  # handles origin tuple, too
        for mode in ['reflect', 'nearest', 'mirror', 'wrap', 'constant']:
            kwargs = dict(axes=axes_size[0], size=axes_size[1], origin=origin, mode=mode)
            ref = generic_filter(input, np.mean, **kwargs)
            res = _generic_filter(input, np.mean, **kwargs)
            np.testing.assert_allclose(res, ref, atol=1e-15)

            kwargs.pop('size')
            kwargs['footprint'] = rng.random(size=axes_size[1] or input.shape) > 0.5
            ref = generic_filter(input, np.mean, **kwargs)
            res = _generic_filter(input, np.mean, **kwargs)
            np.testing.assert_allclose(res, ref, atol=1e-15)
# all passes
```

</details>

However, for decent performance with the code as-written, the user callable has to be vectorized. It is likely safe to just check for the presence of an `axis` keyword argument, but we could also have the user opt-in.