@rgommers 2025-02-23
Interesting, thanks for working on this @mdhaber!

First thought: this might be better as a new function, given that the input function has to be written differently, performance characteristics are going to be very different, and it looks like one or more new keyword arguments may be needed in order to control the memory usage and/or indicate that `function` is vectorized. I'd prototype it as a separate function and benchmark performance vs. memory usage, since that seems the critical bit here. I think small `size`/`footprint` calls are common and worth optimizing for. I'm a bit worried though that it's typical to optimize this by scanning over sizes, and this scales quadratically so on common image sizes like 1024x1024 or 1920x1080 one will run out of memory somewhere in the `size=10-30` range probably.

@jni, @stefanv, @grlee77 you may have more thoughts on this?