@mdhaber 2025-02-23
Limiting memory to 1 GB, here are results of performing a median filter on `uint8` data. Rows are image sizes, columns are filter sizes. Numbers greater than 1 indicated that `vectorized_filter` was faster. Custom `median` callable was implemented with `np.partition` (because `np.median` generally returns the average of two values for even sample sizes and doesn't preserve dtype).

<details>

```python3
import numpy as np
from scipy import ndimage
from timeit import timeit

image_sizes = 2**np.arange(8, 11)
footprint_sizes = 2**np.arange(1, 8)

def median(input, axis=(-1,)):
    input = np.reshape(input, input.shape[:-len(axis)] + (-1,))
    k = input.shape[-1] // 2
    temp = np.partition(input, k)
    return temp[..., k]

rng = np.random.default_rng(238549234965823456)
dtype = np.uint8

times_vf = np.zeros((image_sizes.size, footprint_sizes.size))
times_mf = np.zeros((image_sizes.size, footprint_sizes.size))
times_gf = np.zeros((image_sizes.size, footprint_sizes.size))

for i, image_size in enumerate(image_sizes):
    input = rng.integers(0, 256, size=(image_size, image_size)).astype(dtype)

    for j, footprint_size in enumerate(footprint_sizes):
        print(image_size, footprint_size)

        # res = ndimage.vectorized_filter(input, median, size=footprint_size)
        # ref = ndimage.median_filter(input, size=footprint_size)
        # ref2 = ndimage.generic_filter(input, median, size=footprint_size)
        # np.testing.assert_equal(res, ref)

        times_vf[i, j] = timeit(lambda: ndimage.vectorized_filter(input, median, size=footprint_size), number=1)
        times_mf[i, j] = timeit(lambda: ndimage.median_filter(input, size=footprint_size), number=1)
        times_gf[i, j] = timeit(lambda: ndimage.generic_filter(input, median, size=footprint_size), number=1)

speedup = times_gf/times_vf
print(speedup)
```

</details>

Speedup relative to `median_filter`:
|             | 2 x 2    | 4 x 4    | 8 x 8    | 16 x 16  | 32 x 32  | 64 x 64  | 128 x 128 |
|-------------|----------|----------|----------|----------|----------|----------|-----------|
| 256 x 256   | 0.933844 | 1.180208 | 1.22228  | 1.258281 | 1.28462  | 1.331541 | 1.604247  |
| 512 x 512   | 0.949615 | 1.212079 | 1.248887 | 1.266894 | 1.280168 | 1.279307 | 1.352838  |
| 1024 x 1024 | 0.966779 | 1.217347 | 1.251129 | 1.27092  | 1.288674 | 1.270469 | 1.289567  |


Speedup relative to `generic_filter`:

|             | 2 x 2    | 4 x 4    | 8 x 8    | 16 x 16  | 32 x 32  | 64 x 64  | 128 x 128 |
|-------------|----------|----------|----------|----------|----------|----------|-----------|
| 256 x 256   | 128.3204 | 23.14959 | 7.685969 | 2.774345 | 1.156269 | 0.736984 | 0.859128  |
| 512 x 512   | 137.1367 | 23.79734 | 7.897406 | 2.742188 | 1.138205 | 0.670418 | 0.614689  |
| 1024 x 1024 | 139.6695 | 23.66734 | 7.883612 | 2.785707 | 1.17085  | 0.674698 | 0.547584  |

Interesting that with this filter function, it's always as fast (essentially) or faster than `median_filter`, and `generic_filter` is faster than both for large filter sizes.

---

Using np.mean with `float32` data, here's the speedup relative to `generic_filter`.

|             | 2 x 2    | 4 x 4    | 8 x 8    | 16 x 16  | 32 x 32  | 64 x 64  | 128 x 128 |
|-------------|----------|----------|----------|----------|----------|----------|-----------|
| 256 x 256   | 205.5262 | 113.2836 | 57.42008 | 26.79971 | 14.58288 | 10.18142 | 16.33123  |
| 512 x 512   | 221.4206 | 118.2336 | 60.20738 | 28.20583 | 14.46404 | 9.063608 | 9.115325  |
| 1024 x 1024 | 204.6833 | 120.5235 | 59.14981 | 28.34652 | 14.46693 | 8.567006 | 7.269432  |