@jni 2025-02-24
Super cool! I think most applications of generic_filter are definitely going to be in the 32x32 regime and under (would be ~(10, 10, 10) in 3D), so this will be a very useful function!

As a complete aside, I'm so used to black-formatted, heavily typed code now that seeing this old-school style Python is such a breath of fresh air! 😍😂