@jni 2025-02-24
btw, I don't have bandwidth for a review/deep dive right now, but it looks like you're on the right track re np.pad and the issues that can be fixed with this approach.