@mdhaber 2025-02-24
Thanks @jni! By using `np.pad`, I was able to make some pretty substantial simplifications (beyond just eliminating the custom padding).

This brings up a question, though - `pad` offers all the `mode` options that our filters do and several more. There have been requests for new `mode` options, so shall we just allow all the other padding `mode`s provided by `pad`? To avoid having a separate `reflect_type` option, some of the options would get `_odd` and `_even` variants.

<details>
One complication is that some of the `pad` options have different names from our equivalents, and there is even disagreement about what `reflect` means: `{'nearest': 'edge', 'reflect': 'symmetric', 'mirror': 'reflect'}`. 

Would you prefer just adding the new (distinct) options offered by `pad` without the aliases? 
Or - since this is a new function - use only names accepted by `np.pad`?
If not the latter, there is also a hybrid: accept all the names currently accepted by `ndimage` filters *and* the names accepted by `pad` *except `reflect`* to eliminate ambiguity w.r.t. `symmetric` and `mirror`.
</details>

---

Another question that is a bit more important: the existing `generic_filter` always returns an array with the same type as the input. Should we keep this convention, or (preferably, IMO) respect the output type of the callable? I prefer the latter so that information is preserved by default, and the user can opt-in to losing information (e.g. by providing `output` or using `astype` on the result).

In order to add support for gh-5137, we're already going to have to check the output dimensionality of the callable before allocating the result array (if `output` is not provided), so it's no additional complication to check the dtype.