@jni 2025-02-25
> so shall we just allow all the other padding modes provided by pad?

This is super tricky, but my gut feeling is, do as scipy does (ie use the same terminology as the rest of ndimage, translating to pad as needed — there is probably already a utility function for this?), and leave the harmonising of scipy border modes and np.pad border modes to later generations. 😂 

> Should we keep this convention, or (preferably, IMO) respect the output type of the callable?

I absolutely think it should be the output type of the callable. Passing an output array or also adding a `dtype=` kwarg are both good workarounds for folks that want to preserve the dtype of the input. (.astype involves a copy so I prefer dtype kwargs, which are also more convenient than providing an output array... Though as long as the array is always returned, I guess it's not too bad to do `vectorized_filter(..., out=np.empty_like(image))`.)